pyproject.toml
[tool.poetry]
name = "imagetotext"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "3.0.3"
pillow = "10.3.0"
pytesseract = "0.3.10"
docx = "^0.2.4"
opencv-python = "^4.10.0.82"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
